url changes

+++ deployment/agentMaster.py
@@ -21,7 +21,12 @@ from dateutil import parser
 #serverUrl = 'http://localhost:8888'
 serverUrl = 'https://agentdata-dot-insightfindergae.appspot.com'
 
-def sshInstall(retry,hostname):
+def get_ip_address():
+    s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+    s.connect(("8.8.8.8", 80))
+    return s.getsockname()[0]
+
+def sshInstall(retry,hostname,hostMap):
     global user
     global password
     global userInsightfinder
@@ -36,10 +41,12 @@ def sshInstall(retry,hostname):
         print "Install Fail in", hostname
         q.task_done()
         return
+
     print "Start installing agent in", hostname, "..."
     try:
         s = paramiko.SSHClient()
         s.set_missing_host_key_policy(paramiko.AutoAddPolicy())
+
         if os.path.isfile(password) == True:
             s.connect(hostname, username=user, key_filename = password, timeout=60)
         else:
@@ -48,33 +55,48 @@ def sshInstall(retry,hostname):
         session = transport.open_session()
         session.set_combine_stderr(True)
         session.get_pty()
-        command = "sudo rm -rf insightagent* InsightAgent* && \
-        wget --no-check-certificate https://github.com/insightfinder/InsightAgent/archive/master.tar.gz -O insightagent.tar.gz && \
-        tar xzvf insightagent.tar.gz && \
-        cd InsightAgent-master && deployment/checkpackages.sh && \
-        deployment/install.sh "+ arguments + "\n"
-        #print 'Command: ', command
+        command = "[ -f /etc/cron.d/ifagent ] && echo 'File exist' || echo 'File does not exist'"
         session.exec_command(command)
-        stdin = session.makefile('wb', -1)
         stdout = session.makefile('rb', -1)
-        stdin.write(password+'\n')
-        stdin.flush()
-        session.recv_exit_status() #wait for exec_command to finish
-        s.close()
-        print "Install Succeed in", hostname
-        q.task_done()
-        return
+        if stdout.read().strip(' \t\n\r') == 'File exist':
+            s.close()
+            hostMap[hostname] = 0
+            print "Installation stopped in ", hostname
+            q.task_done()
+            return
+        else:
+            hostMap[hostname] = 1
+            session = transport.open_session()
+            session.set_combine_stderr(True)
+            session.get_pty()
+            nextcommand = "sudo rm -rf insightagent* InsightAgent* && \
+            wget --no-check-certificate https://github.com/amurark/InsightAgent/archive/master.tar.gz -O insightagent.tar.gz && \
+            tar xzvf insightagent.tar.gz && \
+            cd InsightAgent-master && deployment/checkpackages.sh && \
+            deployment/install.sh "+ arguments + "\n"
+            #print 'Command: ', command
+            session.exec_command(nextcommand)
+            stdin1 = session.makefile('wb', -1)
+            stdout1 = session.makefile('rb', -1)
+            stdin1.write(password+'\n')
+            stdin1.flush()
+            session.recv_exit_status() #wait for exec_command to finish
+            s.close()
+            print "Install Succeed in", hostname
+            q.task_done()
+            return
     except paramiko.SSHException, e:
         print "Invalid Username/Password for %s:"%hostname , e
-        return sshInstall(retry-1,hostname)
+        return sshInstall(retry-1,hostname,hostMap)
     except paramiko.AuthenticationException:
         print "Authentication failed for some reason in %s:"%hostname
-        return sshInstall(retry-1,hostname)
+        return sshInstall(retry-1,hostname,hostMap)
     except socket.error, e:
         print "Socket connection failed in %s:"%hostname, e
-        return sshInstall(retry-1,hostname)
+        return sshInstall(retry-1,hostname,hostMap)
     except:
         print "Unexpected error in %s:"%hostname
+        return sshInstall(retry-1,hostname,hostMap)
 
 def sshInstallHypervisor(retry,hostname):
     global user
@@ -184,29 +206,52 @@ if __name__ == '__main__':
     global agentType
     global projectName
     global newInstances
+    #File containing list of instances the agent is installed on. The file doesn't exit the first time the agent runs.
     jsonFile="instancesMetaData.json"
+    #File containing blacklisted instances.
     excludeListFile="excludeList.csv"
     user, projectName, userInsightfinder, licenseKey, samplingInterval, reportingInterval, agentType, password = get_args()
     q = Queue.Queue()
     newInstances = []
     try:
+        #Data to be sent with the POST request to fetch list of instances associated with the project.
         dataString = "projectName="+projectName+"&userName="+userInsightfinder+"&licenseKey="+licenseKey
+        #URL for POST request
         url = serverUrl + "/api/v1/instanceMetaData"
         print "url", url
         print "dataString", dataString
+        #Subprocess to run the CURL command for POST reqest.
         proc = subprocess.Popen(['curl --data \''+ dataString +'\' '+url], stdout=subprocess.PIPE, shell=True)
+        #Get the ouput data and err data.
         (out,err) = proc.communicate()
         print 'Output', out
-        output = json.loads(out)   
+        output = json.loads(out)
+        #If the output doesn't create success field, exit because of error in POST.
         if not output["success"]:
             print "Error: " + output["message"]
             sys.exit(output["message"])
+
+        #Get the dictionary of instances from the output.
         instances = output["data"]
-        newInstances = {} 
+        #This will hold the list of allowed instances, by eliminating the existing instances and the blacklisted instances.
+        allowed_instances = {}
+        self_ip = get_ip_address()
+        #If the instance has the flag 'AutoDeployInsightAgent' set to False, then ignore the instance for auto agent install.
+        for key in instances:
+            if 'AutoDeployInsightAgent' in instances[key] and instances[key]['AutoDeployInsightAgent'] == False:
+                continue
+            else:
+                if str('privateip' in instances[key] and instances[key]['privateip']) == str(self_ip):
+                    continue
+                else:
+                    allowed_instances[key] = instances[key]
+        newInstances = {}
         oldInstances = {}
         excludeList = []
         newKeys = []
         print "Checking Path: ",os.path.join(os.getcwd(),excludeListFile)
+
+        #Also, check the exclude list for blacklisted instances and create a list of excluded instances.
         if os.path.exists(os.path.join(os.getcwd(),excludeListFile)):
             with open(os.path.join(os.getcwd(),excludeListFile), 'rb') as f:
                 reader = csv.reader(f)
@@ -215,47 +260,57 @@ if __name__ == '__main__':
                     print "Added Exclude List: ", row[0]
 
         print "Checking Path: ",os.path.join(os.getcwd(),jsonFile)
+        #If the file doesn't exist, then all allowed instances should be included for agent installation
         if not os.path.exists(os.path.join(os.getcwd(),jsonFile)):
-            newInstances = instances
-            newKeys = instances.keys()
+            newInstances = allowed_instances
+            newKeys = allowed_instances.keys()
+        #If the file exists, remove the instances which already exist in the file.
         else:
             oldInstances = json.load(open(os.path.join(os.getcwd(),jsonFile), "rb" ))
-            newKeys = set(instances.keys()) - set(oldInstances.keys())
-            for instance in instances:
+            #Get the final list of instances.
+            newKeys = set(allowed_instances.keys()) - set(oldInstances.keys())
+            #If the instance is in oldInstances list then, check the time difference between them, if the instances are more than 1 day old then put them in the newlist again.
+            """
+            for instance in allowed_instances:
                 if instance in oldInstances:
                     print "Instance", instance
                     print "DateTimeDiff: ", datetime.now()-parser.parse(oldInstances[instance]["lastSeen"])
                     if (datetime.now()-parser.parse(oldInstances[instance]["lastSeen"])).days > 0:
                         newInstances[instance]=oldInstances[instance]
-                        newInstances[instance]["lastSeen"]=str(datetime.now())  
-                              
+                        newInstances[instance]["lastSeen"]=str(datetime.now())
+            """
         for newKey in newKeys:
-            newInstances[newKey]=instances[newKey]
+            newInstances[newKey]=allowed_instances[newKey]
+            newInstances[newKey]["agentExist"] = False
+            """
             newInstances[newKey]["lastSeen"]=str(datetime.now())
+            """
 
-
-
-                
-        print 'New Instances: ', newInstances   
-        updatedInstances = oldInstances.copy()
-        updatedInstances.update(newInstances) 
-        json.dump(updatedInstances,open(os.path.join(os.getcwd(),jsonFile), "wb" ))
-        
+        #Deploy on all the instances except the blacklisted ones.
         for instanceKey in newInstances:
-                host = newInstances[instanceKey]["publicIp"]
+                host = newInstances[instanceKey]["privateip"]#Changed from publicIp
                 if host in excludeList:
-                    continue;
-                q.put(host)
-        
+                    continue
+                print host
+                #q.put(host)
+        hostMap = {}
         while q.empty() != True:
             host = q.get()
             if agentType == "hypervisor":
                 t = threading.Thread(target=sshInstallHypervisor, args=(3,host,))
             else:
-                t = threading.Thread(target=sshInstall, args=(3,host,))
+                t = threading.Thread(target=sshInstall, args=(3,host,hostMap))
             t.daemon = True
             t.start()
         q.join()
+        #Update the json file, if an agent exists on certain host machines, then update flag accordingly.
+        updatedInstances = oldInstances.copy()
+        updatedInstances.update(newInstances)
+        for key in updatedInstances.keys():
+            ip = str(updatedInstances[key]['privateip'])
+            if ip in hostMap and hostMap[ip] == 0:
+                updatedInstances[key]['agentExist'] = True
+        json.dump(updatedInstances,open(os.path.join(os.getcwd(),jsonFile), "wb" ))
 
     except (KeyboardInterrupt, SystemExit):
         print "Keyboard Interrupt!!"
